refactor(quarkifier): clean architecture with subpackages and god class split
