+++ quarkifier/src/main/java/com/clementguillot/quarkifier/maven/MavenCoordinateParser.java
@@ -1,6 +1,7 @@
-package com.clementguillot.quarkifier;
+package com.clementguillot.quarkifier.maven;
 
 import java.nio.file.Path;
+import java.util.Arrays;
 import java.util.Set;
 
 /**
@@ -11,11 +12,9 @@ import java.util.Set;
  */
 public final class MavenCoordinateParser {
 
-  /** Path segments that mark the boundary before Maven groupId segments begin. */
   private static final Set<String> STOP_SEGMENTS =
       Set.of("external", "v1", "https", "file", "bin", "repo", "jars");
 
-  /** Prefixes that mark non-groupId segments (Bazel/OS roots). */
   private static final String[] STOP_PREFIXES = {"maven", "bazel-", "darwin", "linux", "windows"};
 
   private MavenCoordinateParser() {}
@@ -32,8 +31,6 @@ public final class MavenCoordinateParser {
    * @return parsed coordinates, or a fallback based on the filename
    */
   public static Coordinates parse(Path jarPath) {
-    // Try to resolve symlinks to get the original path (e.g., Coursier cache)
-    // which has the full Maven directory structure for groupId extraction.
     Path resolvedPath = jarPath;
     try {
       resolvedPath = jarPath.toRealPath();
@@ -66,11 +63,17 @@ public final class MavenCoordinateParser {
       return fallback(jarPath);
     }
 
-    return new Coordinates(String.join(".", java.util.Arrays.copyOfRange(parts, groupIdStart, groupIdEnd + 1)), artifactId, version);
+    return new Coordinates(
+        String.join(".", Arrays.copyOfRange(parts, groupIdStart, groupIdEnd + 1)),
+        artifactId,
+        version);
   }
 
   private static boolean isStopSegment(String segment) {
-    if (segment.isEmpty() || segment.contains("+") || segment.contains("=") || segment.contains(".")) {
+    if (segment.isEmpty()
+        || segment.contains("+")
+        || segment.contains("=")
+        || segment.contains(".")) {
       return true;
     }
     if (STOP_SEGMENTS.contains(segment)) return true;
@@ -89,8 +92,6 @@ public final class MavenCoordinateParser {
       name = name.substring(0, name.length() - 4);
     }
 
-    // Try to extract artifactId and version from "artifactId-version" pattern.
-    // Version typically starts with a digit, so find the last "-" followed by a digit.
     int versionSep = -1;
     for (int i = name.length() - 1; i > 0; i--) {
       if (name.charAt(i - 1) == '-' && Character.isDigit(name.charAt(i))) {
