+++ quarkifier/src/test/java/com/clementguillot/quarkifier/extension/MissingDeploymentArtifactExceptionTest.java
@@ -1,4 +1,4 @@
-package com.clementguillot.quarkifier;
+package com.clementguillot.quarkifier.extension;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -6,14 +6,6 @@ import org.junit.jupiter.api.Test;
 
 /** Unit tests for {@link MissingDeploymentArtifactException}. */
 class MissingDeploymentArtifactExceptionTest {
-
-  @Test
-  void getters_returnConstructorValues() {
-    var ex = new MissingDeploymentArtifactException("quarkus-arc-deployment", "quarkus-arc");
-
-    assertEquals("quarkus-arc-deployment", ex.getMissingDeploymentArtifactId());
-    assertEquals("quarkus-arc", ex.getOriginatingExtensionArtifactId());
-  }
 
   @Test
   void message_containsBothIds() {
